Rename niTextureFormatPrefs to textureFormatPrefs to be consistent with the other tables
