+++ misc/package/Data Files/MWSE/core/lib/ni/init.lua
@@ -2,8 +2,6 @@ local ni = {}
 
 ni.animationKeyType = require("ni.animationKeyType")
 ni.gravityForceType = require("ni.gravityForceType")
-ni.niTextureFormatPrefsMipFlag = require("ni.niTextureFormatPrefsMipFlag")
-ni.niTextureFormatPrefsPixelLayout = require("ni.niTextureFormatPrefsPixelLayout")
 ni.particleBombDecayType = require("ni.particleBombDecayType")
 ni.particleBombSymmetryType = require("ni.particleBombSymmetryType")
 ni.pickCoordinateType = require("ni.pickCoordinateType")
@@ -12,6 +10,8 @@ ni.pickSortType = require("ni.pickSortType")
 ni.pickType = require("ni.pickType")
 ni.propertyType = require("ni.propertyType")
 ni.textureFormatPrefsAlphaFormat = require("ni.textureFormatPrefsAlphaFormat")
+ni.textureFormatPrefsMipFlag = require("ni.textureFormatPrefsMipFlag")
+ni.textureFormatPrefsPixelLayout = require("ni.textureFormatPrefsPixelLayout")
 ni.texturingPropertyApplyMode = require("ni.texturingPropertyApplyMode")
 ni.texturingPropertyClampMode = require("ni.texturingPropertyClampMode")
 ni.texturingPropertyFilterMode = require("ni.texturingPropertyFilterMode")
