+++ misc/package/Data Files/MWSE/core/lib/ni/textureFormatPrefsMipFlag.lua
