+++ misc/package/Data Files/MWSE/core/lib/ni/textureFormatPrefsPixelLayout.lua
